Some work on model and animation configs.
git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@205 edc67d0a-ffea-dc11-983e-003048321218

+++ src/java/com/threerings/editor/swing/BasePropertyEditor.java
@@ -183,9 +183,9 @@ public abstract class BasePropertyEditor extends JPanel
     /**
      * Returns a background color darkened by the specified number of shades.
      */
-    protected static Color getDarkerBackground (int shades)
+    protected static Color getDarkerBackground (float shades)
     {
-        int value = BASE_BACKGROUND - shades*SHADE_DECREMENT;
+        int value = BASE_BACKGROUND - (int)(shades*SHADE_DECREMENT);
         return new Color(value, value, value);
     }
